refactor(auth): extract DTOs to dedicated types.rs
