+++ server/src/openapi.rs
@@ -333,23 +333,23 @@ use utoipa::{Modify, OpenApi};
         // Health
         crate::api::HealthResponse,
         // Auth
-        crate::auth::handlers::RegisterRequest,
-        crate::auth::handlers::LoginRequest,
-        crate::auth::handlers::RefreshRequest,
-        crate::auth::handlers::LogoutRequest,
-        crate::auth::handlers::AuthResponse,
-        crate::auth::handlers::UserProfile,
-        crate::auth::handlers::MfaSetupResponse,
-        crate::auth::handlers::MfaBackupCodesResponse,
-        crate::auth::handlers::MfaBackupCodeCountResponse,
-        crate::auth::handlers::MfaVerifyRequest,
-        crate::auth::handlers::UpdateProfileRequest,
-        crate::auth::handlers::UpdateProfileResponse,
-        crate::auth::handlers::ForgotPasswordRequest,
-        crate::auth::handlers::ResetPasswordRequest,
-        crate::auth::handlers::SessionInfo,
-        crate::auth::handlers::SessionListResponse,
-        crate::auth::handlers::RevokeAllResponse,
+        crate::auth::types::RegisterRequest,
+        crate::auth::types::LoginRequest,
+        crate::auth::types::RefreshRequest,
+        crate::auth::types::LogoutRequest,
+        crate::auth::types::AuthResponse,
+        crate::auth::types::UserProfile,
+        crate::auth::types::MfaSetupResponse,
+        crate::auth::types::MfaBackupCodesResponse,
+        crate::auth::types::MfaBackupCodeCountResponse,
+        crate::auth::types::MfaVerifyRequest,
+        crate::auth::types::UpdateProfileRequest,
+        crate::auth::types::UpdateProfileResponse,
+        crate::auth::types::ForgotPasswordRequest,
+        crate::auth::types::ResetPasswordRequest,
+        crate::auth::types::SessionInfo,
+        crate::auth::types::SessionListResponse,
+        crate::auth::types::RevokeAllResponse,
         crate::auth::error::ErrorResponse,
         // DB Models
         crate::db::AuthMethod,
